Fix linting errors in truck check onboarding wizard
- Move VEHICLE_TEMPLATES to separate vehicleTemplates.ts file to fix react-refresh/only-export-components error (line 42)
- Replace sr-only legend with visible text to fix jsx-a11y/label-has-associated-control warning (line 417)
- Add proper styling for legend element in both light and dark modes

CI gate: frontend lint now passes
