+++ frontend/src/features/auth/TruckCheckOnboardingWizard.css
@@ -161,6 +161,15 @@
   margin: 0;
 }
 
+.vehicles-grid > legend {
+  grid-column: 1 / -1;
+  font-size: 0.95rem;
+  font-weight: 500;
+  color: #1a1a1a;
+  margin-bottom: 0.5rem;
+  padding: 0;
+}
+
 @media (min-width: 600px) {
   .vehicles-grid {
     grid-template-columns: 1fr 1fr;
@@ -407,6 +416,10 @@
 @media (prefers-color-scheme: dark) {
   .onboarding-card {
     background: #2a2a2a;
+    color: #e0e0e0;
+  }
+
+  .vehicles-grid > legend {
     color: #e0e0e0;
   }
