+++ frontend/src/features/auth/TruckCheckOnboardingWizard.tsx
@@ -16,6 +16,7 @@ import { useToast } from '../../hooks/useToast';
 import { api } from '../../services/api';
 import { PageTransition } from '../../components/PageTransition';
 import { markTruckCheckOnboardingComplete } from '../../utils/onboardingUtils';
+import { VEHICLE_TEMPLATES } from './vehicleTemplates';
 import './TruckCheckOnboardingWizard.css';
 
 // NSW RFS & partner agencies
@@ -38,134 +39,6 @@ const JURISDICTIONS = [
   { id: 'ACT', label: 'Australian Capital Territory (ACT)' },
 ];
 
-// Standard vehicle type templates available for initial setup
-export const VEHICLE_TEMPLATES = [
-  // NSW RFS
-  {
-    id: 'std-cat1-tanker',
-    label: 'Cat 1 Tanker',
-    description: 'NSW RFS Category 1 tanker for rural fire suppression',
-    agency: 'NSW RFS',
-  },
-  {
-    id: 'std-cat7-tanker',
-    label: 'Cat 7 Tanker',
-    description: 'NSW RFS Category 7 tanker for rural fire suppression',
-    agency: 'NSW RFS',
-  },
-  {
-    id: 'std-cat9-tanker',
-    label: 'Cat 9 Tanker (Light)',
-    description: 'NSW RFS Category 9 light tanker for rapid response',
-    agency: 'NSW RFS',
-  },
-  {
-    id: 'std-urban-pumper',
-    label: 'Urban Pumper',
-    description: 'NSW RFS urban structure fire pumper (Cat 10/11)',
-    agency: 'NSW RFS',
-  },
-  {
-    id: 'std-group-personnel-vehicle',
-    label: 'Group / Personnel Vehicle',
-    description: 'NSW RFS group transport and personnel vehicle',
-    agency: 'NSW RFS',
-  },
-  // Fire & Rescue NSW
-  {
-    id: 'std-class3-pumper',
-    label: 'Class 3 Pumper',
-    description: 'Fire & Rescue NSW Class 3 structure fire pumper',
-    agency: 'Fire & Rescue NSW',
-  },
-  {
-    id: 'std-rescue-pumper',
-    label: 'Rescue Pumper',
-    description: 'Fire & Rescue NSW multi-purpose rescue and pumping vehicle',
-    agency: 'Fire & Rescue NSW',
-  },
-  {
-    id: 'std-aerial-ladder',
-    label: 'Aerial Ladder Platform',
-    description: 'Fire & Rescue NSW aerial ladder platform for high-rise operations',
-    agency: 'Fire & Rescue NSW',
-  },
-  {
-    id: 'std-hazmat-appliance',
-    label: 'Hazmat Appliance',
-    description: 'Fire & Rescue NSW hazardous materials response vehicle',
-    agency: 'Fire & Rescue NSW',
-  },
-  // NSW SES
-  {
-    id: 'std-general-purpose-vehicle',
-    label: 'General Purpose Vehicle',
-    description: 'NSW SES general purpose support and transport vehicle',
-    agency: 'NSW SES',
-  },
-  {
-    id: 'std-light-rescue',
-    label: 'Light Rescue Vehicle',
-    description: 'NSW SES light rescue response vehicle',
-    agency: 'NSW SES',
-  },
-  {
-    id: 'std-heavy-rescue',
-    label: 'Heavy Rescue Vehicle',
-    description: 'NSW SES heavy rescue and extrication vehicle',
-    agency: 'NSW SES',
-  },
-  {
-    id: 'std-high-clearance-vehicle',
-    label: 'High Clearance Vehicle',
-    description: 'NSW SES high clearance terrain vehicle',
-    agency: 'NSW SES',
-  },
-  {
-    id: 'std-storm-truck',
-    label: 'Storm Truck',
-    description: 'NSW SES storm response and cleanup vehicle',
-    agency: 'NSW SES',
-  },
-  // Marine Rescue NSW
-  {
-    id: 'std-offshore-rescue-vessel',
-    label: 'Offshore Rescue Vessel',
-    description: 'Marine Rescue NSW offshore rescue vessel',
-    agency: 'Marine Rescue NSW',
-  },
-  {
-    id: 'std-inshore-rescue-vessel',
-    label: 'Inshore Rescue Vessel',
-    description: 'Marine Rescue NSW inshore rescue vessel',
-    agency: 'Marine Rescue NSW',
-  },
-  {
-    id: 'std-rescue-water-craft',
-    label: 'Rescue Water Craft',
-    description: 'Marine Rescue NSW rescue water craft',
-    agency: 'Marine Rescue NSW',
-  },
-  // Multi-agency
-  {
-    id: 'std-command-vehicle',
-    label: 'Command Vehicle',
-    description: 'Inter-agency command and coordination vehicle',
-    agency: 'Multi-agency',
-  },
-  {
-    id: 'std-crew-transport',
-    label: 'Crew Transport',
-    description: 'Personnel transport and support vehicle',
-    agency: 'Multi-agency',
-  },
-  {
-    id: 'std-logistics-trailer',
-    label: 'Logistics Trailer',
-    description: 'Mobile logistics and supply support trailer',
-    agency: 'Multi-agency',
-  },
-];
 
 type Step = 'jurisdiction' | 'agency' | 'vehicles' | 'confirm';
 
@@ -412,7 +285,7 @@ export function TruckCheckOnboardingWizard() {
                 )}
 
                 <fieldset className="vehicles-grid">
-                  <legend className="sr-only">Select your vehicles</legend>
+                  <legend>Which vehicles do you operate? (select at least one)</legend>
                   {VEHICLE_TEMPLATES.map((vehicle) => (
                     <label key={vehicle.id} className="vehicle-option">
                       <input
